+++ batik-gvt/src/main/java/org/apache/batik/gvt/text/ArabicTextHandler.java
@@ -453,7 +453,7 @@ public class ArabicTextHandler {
                         && nextForm.equals(ARABIC_INITIAL)) {
                         // look for an isolated ligature
                         int substChar = ArabicTextHandler.getSubstituteChar
-                            (c, nextChar,ARABIC_ISOLATED.intValue());
+                            (c, nextChar, ARABIC_ISOLATED);
                         if (substChar > -1) {
                             substString.append((char)substChar);
                             i++;
@@ -462,7 +462,7 @@ public class ArabicTextHandler {
                     } else if (form.equals(ARABIC_TERMINAL)) {
                         // look for a terminal ligature
                         int substChar = ArabicTextHandler.getSubstituteChar
-                            (c, nextChar,ARABIC_TERMINAL.intValue());
+                            (c, nextChar, ARABIC_TERMINAL);
                         if (substChar > -1) {
                             substString.append((char)substChar);
                             i++;
@@ -472,7 +472,7 @@ public class ArabicTextHandler {
                                && nextForm.equals(ARABIC_MEDIAL)) {
                         // look for a medial ligature
                         int substChar = ArabicTextHandler.getSubstituteChar
-                            (c, nextChar,ARABIC_MEDIAL.intValue());
+                            (c, nextChar, ARABIC_MEDIAL);
                         if (substChar > -1) {
                             substString.append((char)substChar);
                             i++;
